+++ R/arma_acf.R
@@ -1,23 +1,27 @@
-#' Calculate the autocorrelation of an `ARMA(p,q)` filter applied to a stationary process.
+#' Calculate the autocorrelation of an ARMA(p,q) filter applied to a stationary process.
 #'
 #' @param acf Vector of length `N + m` of autocorrelations of the original stationary process.
 #' @param phi Vector of `p >= 0` coefficients defining the autoregressive part of the filter (see 'Details').
 #' @param rho Vector of `q >= 0` coefficients defining the moving-average part of the filter (see 'Details').
-#' @param m Order of the MA approximation to the `ARMA(p,q)` filter.
+#' @param m Order of the MA approximation to the ARMA(p,q) filter.
 #' @return A vector of length `N` giving the autocorrelation of the filtered process.
-#' @details Let `X[n]` denote the observation of the original process at time `n`, and `Y[n]` denote the corresponding observation of the filtered process.  The AR(p,q) filter model is defined as
+#' @details Let `X_n` denote the observation of the original process at time `n`, and `Y_n` denote the corresponding observation of the filtered process.  The AR(p,q) filter model is defined as
 #' ```
-#' Y[n] = X[n] + sum_{i=1}^p phi[i] Y[n-i] + sum_{j=1}^q + rho[j] X[n-j].
+#' Y_n = X_n + sum_{i=1}^p phi_i Y_{n-i} + sum_{j=1}^q + rho_j X_{n-j}.
 #' ```
-#' The `MA(m)` approximation is of the form
+#' The autocorrelation of `Y_n` can be computed exactly from that of `X_n` for pure moving-average filters with `p = 0`.  The general ARMA(p,q) filter is first approximated by a moving-average process of order `m` of the form
 #' ```
 #' Y[n] = X[n] + sum_{j=1}^m psi[j] X[n-j],
 #' ```
 #' where the coefficients `psi` are determined using the method of Brockwell & Davis (1991) implemented in [stats::ARMAtoMA()].
+#'
 #' @importFrom fftw FFT IFFT
 #' @importFrom stats ARMAtoMA
+#'
+#' @references Brockwell, P.J. and Davis, R.A. "Time Series: Theory and Methods" (1991).  Springer, New York.  <https://doi.org/10.1007/978-1-4899-0004-3>.
+#'
 #' @export
-arma_acf <- function(acf, phi = numeric(), rho = numeric(), m = 20) {
+arma_acf <- function(acf, phi = numeric(), rho = numeric(), m) {
   ## if(length(rho) == 0) stop("rho must have non-zero length.")
   ## if(length(phi) == 0) {
   ##   psi <- rho
@@ -25,6 +29,7 @@ arma_acf <- function(acf, phi = numeric(), rho = numeric(), m = 20) {
   ##   psi <- arma2ma(phi = phi, rho = rho, q = m)
   ## }
   psi <- ARMAtoMA(ar = phi, ma = rho, lag.max = m)
+  if(length(acf) <= m) stop("acf must be of length greater than m.")
   ma_acf(acf, rho = psi)
 }
 
@@ -34,9 +39,9 @@ arma_acf <- function(acf, phi = numeric(), rho = numeric(), m = 20) {
 #' @param rho Vector of `q >= 0` coefficients of the `MA(q)` filter (see 'Details').
 #' @return A vector of length `N` giving the autocorrelation of the filtered process.
 #'
-#' @details Let `X[n]` denote the observation of the original process at time `n`, and `Y[n]` denote the corresponding observation of the filtered process.  The `MA(q)` filter model is defined as
+#' @details Let `X_n` denote the observation of the original process at time `n`, and `Y_n` denote the corresponding observation of the filtered process.  The MA(q) filter model is defined as
 #' ```
-#' Y[n] = X[n] + sum_{j=1}^q rho[j] X[n-j].
+#' Y_n = X_n + sum_{j=1}^q rho_j X_{n-j}.
 #' ```
 #' @export
 ma_acf <- function(acf, rho = numeric()) {